Remove inst[] and putLab(), as they're no longer needed.
Don't call bufEnsure(); it's not needed any more because of the change
to *read.c.

Subtract 1 from the length of the array in S(), to omit the final NUL.

In putImm, check r, not FLAG_R (which is spurious, and has now been
removed from translate.h).

Add SP macro to put spaces in the output, and use it.

Add call to Lab() to wrLab(): the label name needs to be written out
too.