Simplify the code and restructure for greater efficiency; in particular, remove hashFind, instead exposing hashSearch as hashGet; hashInsert becomes hashGet, which takes a HashLink (presumably returned by a previous hashGet). This removes the need to search the hash table twice when inserting a new item, and checking first to see if it already exists (which Mite needs to do).
This also has the consequence of needing to expose HashLink in the
header file.

The code is now rather more specialised to its application, but the
slightly less neat interface is compensated for by the greater
efficiency, with an eye to using this code for symbol lookup during
dynamic run-time binding.