Add several TODOs which have nothing to do with Mite; this file should really not be in Mite's CVS repo, and the plan is to merge it with the proposed Lua standard libraries project, if I or anyone else ever gets that off the ground.
Add traceCall for debugging based on lua/test/trace-calls.lua, and add
some machinery to recognise a standard "_DEBUG" table which contains
debugging options.

Reorder the second and third parameters of wrap, and make the third
default to the second. This makes more sense in many cases, as the
majority of lines (2-n) are more important than the first.