Make some changes to the structure of Translator as a result of adding interpWrite.lua:
Add a finish field to Translator, so that a reader or writer can add
finalisation code (run after dangle resolution).

Make writeUInt a macro, so that it can be implemented more efficiently
by some writers.

Add comments where code is spliced in to a translator from a Reader or
Writer, so that it's easier to see where it came from.

Remove the ops() function from mkTrans.lua, which was defunct.