Reorganise translator generation to generate one big C file (translators.c). This means that the prelude for each reader and writer is only compiled once.
To make this possible, most macros and routines in readers and writers
are prefixed with the name of the reader or writer followed by R_ or
W_ as appropriate.

To keep writers easier to read and write (!) the writer macros are
*not* prefixed. Instead, they are moved out of the prelude into a new
"macros" section, which first undefs all the macros, then defines
them. This is pasted directly before each translator that uses the
writer.