Fix a number of problems with the object encoding
The main change is to start encodings at 0 rather than 1. I think I used to
start at 1 so that it was easier to be sure that code was valid (and did not
contain “default zeroes”), but this violates the principle of least
surprise (most encodings are zero-based), and also makes validation a little
harder (because zero must be checked as a separate invalid case).

Lower the maximum number of registers allowed from 256 to 128, as otherwise
the maximum number (including special registers) cannot be encoded in a
single byte.

Remove the unused “temporary register, T” from the spec. This was added with
the (currently commented-out) C call instructions, but apparently unrelated
to them. I’m not sure now what it was meant for.

Remove the “removed for future expansion” note about opcodes (which in any
case gave the wrong range). The implication was that other opcodes could be
used by the user. For now, simply leave it unspecified.